- bug fix|
- new seeder for user Model 
- user - bootcamp relation 
- user - course relation

+++ controllers/courses.js
@@ -11,7 +11,8 @@ const Bootcamp = require('../models/Bootcamp');
 
     // GET /api/v1/bootcamps/:bootcampId/courses
     if(req.params.bootcampId) {
-        const courses = Course.find({bootcamp: req.params.bootcampId});
+        const courses = await Course.find({bootcamp: req.params.bootcampId});
+
         return res.status(200).json({
             success: true,
             count : courses.length,
@@ -55,6 +56,7 @@ const Bootcamp = require('../models/Bootcamp');
  exports.addCourse =  asyncHandler(async (req, res, next) => {
 
     req.body.bootcamp = req.params.bootcampId;
+    req.body.user = req.user.id;
 
     const bootcamp = await Bootcamp.findById(req.params.bootcampId);
  
@@ -62,6 +64,14 @@ const Bootcamp = require('../models/Bootcamp');
      return next(
          new ErrorResponse(`No bootcamp with id ${req.params.bootcampId}`, 404)
      );
+    }
+
+    //check if user is the owner of bootcamp and is also has admin role
+    if(bootcamp.user.toString() !== req.user.id && req.user.role !=='admin'){
+        return next(
+            new ErrorResponse(`User ${req.user.id} is not authorized to add a course to bootcamp ${bootcamp._id}`,
+         401)
+        )
     }
 
     const course = await Course.create(req.body)
@@ -88,6 +98,13 @@ const Bootcamp = require('../models/Bootcamp');
      );
     }
 
+    if(course.user.toString() != req.user.id && req.user.role !== 'admin'){
+        return next(
+            new ErrorResponse(`User ${req.user.id} is not authorized to update course ${course._id}`,
+         401)
+        )
+    }
+
     course = await Course.findByIdAndUpdate(req.params.id, req.body, {
         new : true,
         runValidators: true
@@ -112,6 +129,13 @@ const Bootcamp = require('../models/Bootcamp');
      return next(
          new ErrorResponse(`No course with id ${req.params.id}`, 404)
      );
+    }
+
+    if(course.user.toString() != req.user.id && req.user.role !== 'admin'){
+        return next(
+            new ErrorResponse(`User ${req.user.id} is not authorized to delete course ${course._id}`,
+         401)
+        )
     }
 
     await course.deleteOne();
